+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Add a filter to allow modifying the ActivityPub preview template
 * `@mentions` in the JSON representation of the reply
 * Add settings to enable/disable e-mail notifications for new followers and direct messages
 
 ### Improved
 
+* Direct Messages: Test for the user being in the to field
 * Direct Messages: Improve HTML to e-mail text conversion
 
 ## [4.5.1] - 2024-12-18